+++ RCaNmodel/src/sampleCaNCPP.cpp
@@ -41,7 +41,7 @@ void updateS(Eigen::MatrixXd &S, Eigen::MatrixXd &S2, Eigen::MatrixXd &S0, Eigen
   } else {
     S2=x*x.transpose();
   }
-
+  
 }
 
 
@@ -104,13 +104,13 @@ IntegerVector order_(const NumericVector & x) {
 
 
 List cpgs(const int N, const Eigen::MatrixXd &A ,
-                     const Eigen::VectorXd &b,
-                     const Eigen::VectorXd &x0,
-                     const int thin=1,
-                     const bool gibbs=true,
-                     const int seed=1,
-                     const int stream=1,
-                     Rcpp::Nullable<Eigen::MatrixXd> covMat = R_NilValue) {
+          const Eigen::VectorXd &b,
+          const Eigen::VectorXd &x0,
+          const int thin=1,
+          const bool gibbs=true,
+          const int seed=1,
+          const int stream=1,
+          Rcpp::Nullable<Eigen::MatrixXd> covMat = R_NilValue) {
   dqrng::dqRNGkind("Xoroshiro128+");
   dqrng::dqset_seed(IntegerVector::create(seed),
                     IntegerVector::create(stream));
@@ -118,7 +118,7 @@ List cpgs(const int N, const Eigen::MatrixXd &A ,
   Eigen::ArrayXd critHist=Eigen::ArrayXd::Constant(p,999999); // keep tracks of the last criterion values
   int m=A.rows();
   double inf = std::numeric_limits<double>::max();
-
+  
   // Check input arguments
   if (m < (p+1) || b.size()!=m || x0.size()!=p){
     throw std::range_error("dimensions mismatch");
@@ -128,28 +128,28 @@ List cpgs(const int N, const Eigen::MatrixXd &A ,
   Eigen::MatrixXd x(p,1);
   Eigen::MatrixXd y(p,1);
   x=x0;
-
+  
   // Initialize variables for keeping track of sample mean, covariance
   // and isotropic transform.
   Eigen::MatrixXd M(p,1);
   M.setZero();
   Eigen::MatrixXd S2(p,p);
   S2.setZero();
-
+  
   // outer products.
   Eigen::MatrixXd S(p,p);
   Eigen::MatrixXd S0(p,p);
   S.setIdentity();
-
+  
   IntegerVector index=Rcpp::seq(0,p-1);
   NumericVector u(p);
-
+  
   Eigen::MatrixXd T1(p,p);
   Eigen::MatrixXd T2(p,p);
   T1.setIdentity();
   Eigen::MatrixXd W(m,p);
   Eigen::VectorXd delta(m); //vector used to store distance to bounds
-
+  
   W = A;
   bool adapt=true;
   bool updatingS=true;
@@ -182,8 +182,8 @@ List cpgs(const int N, const Eigen::MatrixXd &A ,
   while (isample<N){               //sampling loop
     //std::random_shuffle(index.begin(), index.end()); //we change the order to
     //limit the influence of initial ordering
-
-      index=dqrng::dqsample_int(p,p,false);
+    
+    index=dqrng::dqsample_int(p,p,false);
     y=x;
     // compute approximate stochastic transformation
     if ((stage==1 && discard==0) || (stage>0 && updatingS==true)){ //first true
[…]
