### 2022-09-30 0ba9994
fix sampleCaN
