+++ RCaNmodel/src/sampleCaNCPP.cpp
@@ -64,7 +64,7 @@ IntegerVector order_(const NumericVector & x) {
 //' @param b a vector of length equals to nrow(A)
 //' @param x0 a vector of length equals to nrcol(A) that should be in the polytope, for example returned by \code{\link{chebyCentre}}
 //' @param thin thinning interval
-//' @param gibbs if true, gibbs sampling, else hitandrun
+//' @param method (1 gibbs, 2 hit-and-run, 3 chrr) 
 //' @param seed seed of the dqrng generator
 //' @param stream stream of the dqrng generator
 //' @param covMat prespecified covmatrix (avoid initialisation and discard) if
@@ -94,7 +94,7 @@ List cpgs(const int N, const Eigen::MatrixXd &A ,
           const Eigen::VectorXd &b,
           const Eigen::VectorXd &x0,
           const int thin=1,
-          const bool gibbs=true,
+          const int method=1,
           const int seed=1,
           const int stream=1,
           Rcpp::Nullable<Eigen::MatrixXd> covMat = R_NilValue) {
@@ -164,6 +164,11 @@ List cpgs(const int N, const Eigen::MatrixXd &A ,
   int runupmax= p*log2(p); //https://doi.org/10.1021/acs.jproteome.5b01029
   int sampleit=0; //total number of iteration during sampling, useful for thin
   int discardmax=runupmax;
+  
+  if (method == 3 ) {//chrr
+
+  }
+
   while (isample<N){               //sampling loop
     //std::random_shuffle(index.begin(), index.end()); //we change the order to
     //limit the influence of initial ordering
@@ -171,7 +176,7 @@ List cpgs(const int N, const Eigen::MatrixXd &A ,
     index=dqrng::dqsample_int(p,p,false);
     y=x;
     // compute approximate stochastic transformation
-    if ((stage==1 && discard==0) || (stage>0 && updatingS==true)){ //first true
+    if (((stage==1 && discard==0) || (stage>0 && updatingS==true)) && method <2){ //first true
       //sample, we now make the isotropic transformation
       ldltOfS.compute(S.transpose());
       D=ldltOfS.vectorD().cwiseMax(Dzero).asDiagonal();
@@ -182,7 +187,7 @@ List cpgs(const int N, const Eigen::MatrixXd &A ,
     }
     if (stage>0) y=T2*y; //otherwise y=I^-1 * y=y
     
-    if (gibbs == true || stage == 0){
+    if (method == 1 || stage == 0){
       NumericVector alea2=dqrng::dqrunif(p);
       // choose p new components
       for (int ip=0;ip<p;++ip){
@@ -213,7 +218,7 @@ List cpgs(const int N, const Eigen::MatrixXd &A ,
         //constraint with the delta of parameter)
       }
     }
-    else {
+    else if (method >=1) { //hit and run
       Eigen::Map<Eigen::VectorXd> u(Rcpp::as<Eigen::Map<Eigen::VectorXd> >(dqrng::dqrnorm(p)));
       u /= u.norm();
       z = W * u;
@@ -238,21 +243,47 @@ List cpgs(const int N, const Eigen::MatrixXd &A ,
     if (stage==0){//still in adapting phase
       ++runup;
       if (updatingS) updateS(S, S2, S0, M, delta0, x, runup);
-      if (runup==runupmax){
+      if (runup>p){
+        stage=1;
+        updatingS=false;
+        Rcout<<"########adapation successful after "<<runup<<" iterations"<<std::endl;
+        discardmax=runup;
+      } else if (runup==runupmax){
         updatingS=false;
         stage=1;
[…]
