+++ RCaNmodel/src/sampleCaNCPP.cpp
@@ -8,6 +8,7 @@
 #include <Eigen/Eigenvalues>
 #include <Eigen/LU>
 #include <testthat.h>
+#include "sampleCaNCPP.h"
 
 
 // [[Rcpp::depends(RcppEigen)]]
@@ -25,7 +26,7 @@ using Eigen::VectorXd;
 
 
 
-void updateS(Eigen::MatrixXd &S, Eigen::MatrixXd &S2, Eigen::MatrixXd &S0, Eigen::MatrixXd &M, Eigen::MatrixXd &delta0, const Eigen::VectorXd &x, int iter){
+void updateS(Eigen::MatrixXd &S, Eigen::MatrixXd &S2, Eigen::MatrixXd &S0, Eigen::VectorXd &M, Eigen::VectorXd &delta0, const Eigen::VectorXd &x, int iter){
   delta0 = x - M; // delta new point wrt old mean
   M+= delta0/(double)iter;     // sample mean
   if (iter>1){
@@ -109,15 +110,17 @@ List cpgs(const int N, const Eigen::MatrixXd &A ,
   if (m < (p+1) || b.size()!=m || x0.size()!=p){
     throw std::range_error("dimensions mismatch");
   }
+  Eigen::VectorXd b2(b.size());
+  b2=b;
   // Initialisation
   Eigen::MatrixXd X(N,p);
-  Eigen::MatrixXd x(p,1);
-  Eigen::MatrixXd y(p,1);
+  Eigen::VectorXd x(p);
+  Eigen::VectorXd y(p);
   x=x0;
   
   // Initialize variables for keeping track of sample mean, covariance
   // and isotropic transform.
-  Eigen::MatrixXd M(p,1);
+  Eigen::VectorXd M(p);
   M.setZero();
   Eigen::MatrixXd S2(p,p);
   S2.setZero();
@@ -139,11 +142,11 @@ List cpgs(const int N, const Eigen::MatrixXd &A ,
   W = A;
   bool adapt=true;
   bool updatingS=true;
-  Eigen::MatrixXd d(m,1);
-  Eigen::MatrixXd d2(m,1);
-  Eigen::MatrixXd delta0(p,1);
-  Eigen::MatrixXd delta1(p,1);
-  Eigen::MatrixXd z(m,1);
+  Eigen::VectorXd d(m);
+  Eigen::VectorXd d2(m);
+  Eigen::VectorXd delta0(p);
+  Eigen::VectorXd delta1(p);
+  Eigen::VectorXd z(m);
   Eigen::MatrixXd L(p,p);
   Eigen::MatrixXd D(p,p);
   Eigen::VectorXd Dtmp(p);
@@ -164,19 +167,23 @@ List cpgs(const int N, const Eigen::MatrixXd &A ,
   int runupmax= p*log2(p); //https://doi.org/10.1021/acs.jproteome.5b01029
   int sampleit=0; //total number of iteration during sampling, useful for thin
   int discardmax=runupmax;
+  MatrixXd T = MatrixXd::Identity(m, m);
+  MatrixXd N_total = MatrixXd::Identity(m, m);
+  MatrixXd N_total_inv(m, m);
+  VectorXd p_shift=VectorXd::Zero(m);
   
   if (method == 3 ) {//chrr
-
+    round(W, b2, x, N_total, p_shift, T, 80);
+    N_total_inv=N_total.inverse();
   }
 
   while (isample<N){               //sampling loop
     //std::random_shuffle(index.begin(), index.end()); //we change the order to
     //limit the influence of initial ordering
     
-    index=dqrng::dqsample_int(p,p,false);
     y=x;
     // compute approximate stochastic transformation
[…]
