+++ RCaNmodel/R/chebyCentre.R
@@ -54,7 +54,7 @@ chebyCentre <- function(A, b, lower = NULL, upper = NULL) {
     upper = c(upper, Inf)
   } else{
     A <- rbind(A, diag(-1, ncol(A)))
-    b <- c(b, rep(0, ncol(A)))
+    b <- c(b, rep(Inf, ncol(A)))
   }
   
   n <- dim(A)[1]
