### 2022-10-05 70eda17
improve sampleCaN
