### 2022-10-05 94e02b5
improve sampleCaN
