+++ RCaNmodel/R/sampleCaN.R
@@ -78,14 +78,14 @@ sampleCaN <- function(myCaNmod,
     `%myinfix%` <- `%dopar%`
   }
   i <- NULL
-  
+  print("##Initializing")
   #we removed parameters that are fixed because of A or almost
   bounds <- getAllBoundsParam(list(A = as.matrix(myCaNmod$A),
                                    b = myCaNmod$b),
-                                   progressBar = FALSE)
+                                   progressBar = TRUE)
   
   #if some are fixed, we add them to C
-  fixed <- which(abs(bounds[, 2] - bounds[, 3]) < 1e-7)
+  fixed <- which(abs(bounds[, 2] - bounds[, 3]) < 1e-5)
   if (length(fixed) > 0){
     myCaNmod$v <- c(myCaNmod$v, rowMeans(bounds[fixed, 2:3, drop = FALSE]))
     newC <- Matrix(0, length(fixed), ncol(myCaNmod$A))
@@ -95,8 +95,8 @@ sampleCaN <- function(myCaNmod,
   
   #now we restric to the degenerate subspace
   if (nrow(myCaNmod$C) > 0){
-    lpmod <- defineLPMod(myCaNmod$A[-seq_len(nrow(myCaNmod$A)), , drop = FALSE],
-                         myCaNmod$b[-seq_len(nrow(myCaNmod$A))], 
+    lpmod <- defineLPMod(myCaNmod$A,
+                         myCaNmod$b, 
                          myCaNmod$C, 
                          myCaNmod$v,
                          maximum = FALSE)
@@ -125,93 +125,68 @@ sampleCaN <- function(myCaNmod,
   #now we presolve the model to simply the polytope
   
   
+  print("##Sampling")
   
   res <- foreach(i = 1:nchain) %myinfix% {
-    find_init <- FALSE
-    nbiter <- 0
-    while (nbiter < 100 & !find_init) {
+
       # lp_model <- defineLPMod(myCaNmod$A, myCaNmod$b, myCaNmod$C, myCaNmod$v,
       #                         maximum = FALSE,
       #                         ob = runif(ncol(myCaNmod$A)))
       presolved <- presolveLPMod(A2,
                                  b2,
+                                 lower = rep(-Inf, ncol(A2)),
+                                 upper = rep(Inf, ncol(A2)),
                                  sense = "min")
       A3 <- presolved$A
       b3 <- presolved$b
       
       
+      x0 <- chebyCenter(A3, 
+                        b3,
+                        lower = presolved$lower,
+                        upper = presolved$upper)
       
+      #to avoid computations problems, we slightly shift b3
+      b3 <- b3 + 1e-7
       #we add the bounds as constraints
-      if (!is.null(presolved$lower)){
-        nonnull <- which(is.finite(presolved$lower))
-        if (length(nonnull) > 0){
-          bounds <- matrix(0, length(nonnull), ncol(A3))
-          bounds[cbind(seq_len(length(nonnull)),
-                       nonnull)] <- - 1
+      nonnull <- which(is.finite(presolved$lower))
+      if (length(nonnull) > 0){
+        bounds <- matrix(0, length(nonnull), ncol(A3))
+        bounds[cbind(seq_len(length(nonnull)),
+                     nonnull)] <- - 1
           A3 <- rbind(A3, bounds)
           b3 <- c(b3, -presolved$lower[nonnull])
         }
-      } else{
-        A3 <- rbind(A3, diag(-1, ncol(A3)))
-        b3 <- c(b3, rep(0, ncol(A3)))
-      }
-      
-      if (!is.null(presolved$upper)){
+     
         nonnull <- which(is.finite(presolved$upper))
         if (length(nonnull) > 0){
[…]
