### 2022-10-06 475ba7b
sampleCaN improved now works for gibbs
