### 2022-10-07 #4804 @fabian-s
:wave: Hi @eafulton, hi @ethanwhite, hi @juanvillada!
Would any of you be willing to review this submission for JOSS?
