+++ RCaNmodel/inst/include/RCaNmodel_RcppExports.h
@@ -25,17 +25,17 @@ namespace RCaNmodel {
         }
     }
 
-    inline List cpgs(const int N, const Eigen::MatrixXd& A, const Eigen::VectorXd& b, const Eigen::VectorXd& x0, const int thin = 1, const int method = 1, const int seed = 1, const int stream = 1, Rcpp::Nullable<Eigen::MatrixXd> covMat = R_NilValue) {
-        typedef SEXP(*Ptr_cpgs)(SEXP,SEXP,SEXP,SEXP,SEXP,SEXP,SEXP,SEXP,SEXP);
+    inline List cpgs(const int N, const Eigen::MatrixXd& A, const Eigen::VectorXd& b, const Eigen::VectorXd& x0, const int thin = 1, const int method = 1, const int seed = 1, const int stream = 1, Rcpp::Nullable<Eigen::MatrixXd> covMat = R_NilValue, Rcpp::Nullable<Eigen::MatrixXd> savedN_total = R_NilValue, Rcpp::Nullable<Eigen::VectorXd> savedp_shift = R_NilValue) {
+        typedef SEXP(*Ptr_cpgs)(SEXP,SEXP,SEXP,SEXP,SEXP,SEXP,SEXP,SEXP,SEXP,SEXP,SEXP);
         static Ptr_cpgs p_cpgs = NULL;
         if (p_cpgs == NULL) {
-            validateSignature("List(*cpgs)(const int,const Eigen::MatrixXd&,const Eigen::VectorXd&,const Eigen::VectorXd&,const int,const int,const int,const int,Rcpp::Nullable<Eigen::MatrixXd>)");
+            validateSignature("List(*cpgs)(const int,const Eigen::MatrixXd&,const Eigen::VectorXd&,const Eigen::VectorXd&,const int,const int,const int,const int,Rcpp::Nullable<Eigen::MatrixXd>,Rcpp::Nullable<Eigen::MatrixXd>,Rcpp::Nullable<Eigen::VectorXd>)");
             p_cpgs = (Ptr_cpgs)R_GetCCallable("RCaNmodel", "_RCaNmodel_cpgs");
         }
         RObject rcpp_result_gen;
         {
             RNGScope RCPP_rngScope_gen;
-            rcpp_result_gen = p_cpgs(Shield<SEXP>(Rcpp::wrap(N)), Shield<SEXP>(Rcpp::wrap(A)), Shield<SEXP>(Rcpp::wrap(b)), Shield<SEXP>(Rcpp::wrap(x0)), Shield<SEXP>(Rcpp::wrap(thin)), Shield<SEXP>(Rcpp::wrap(method)), Shield<SEXP>(Rcpp::wrap(seed)), Shield<SEXP>(Rcpp::wrap(stream)), Shield<SEXP>(Rcpp::wrap(covMat)));
+            rcpp_result_gen = p_cpgs(Shield<SEXP>(Rcpp::wrap(N)), Shield<SEXP>(Rcpp::wrap(A)), Shield<SEXP>(Rcpp::wrap(b)), Shield<SEXP>(Rcpp::wrap(x0)), Shield<SEXP>(Rcpp::wrap(thin)), Shield<SEXP>(Rcpp::wrap(method)), Shield<SEXP>(Rcpp::wrap(seed)), Shield<SEXP>(Rcpp::wrap(stream)), Shield<SEXP>(Rcpp::wrap(covMat)), Shield<SEXP>(Rcpp::wrap(savedN_total)), Shield<SEXP>(Rcpp::wrap(savedp_shift)));
         }
         if (rcpp_result_gen.inherits("interrupted-error"))
[…]
