+++ RCaNmodel/src/sampleCaNCPP.cpp
@@ -75,6 +75,12 @@ IntegerVector order_(const NumericVector & x) {
 //' This function is based on an initial matlab code developped called CPRND
 //' (https://ch.mathworks.com/matlabcentral/fileexchange/34208-uniform-distribution-over-a-convex-polytope)
 //' It generates samples within the complex polytope defined by \eqn{A \cdot x \leqslant   b}
+//' The CHRR algorithm is a C++ translation of cobratoolbox code written
+//' by Yin Zhang licensed under GNU GPL-3 https://github.com/opencobra/cobratoolbox/ 
+//' and of matlab code written by Ben Cousins 
+//' (https://github.com/Bounciness/Volume-and-Sampling) is a C++ translation of
+//' matlab functions provided in the opencobra toolbox and 
+//' in this github repository
 //'
 //' @return a matrix with one row per sample and one column per parameter
 //' @examples
@@ -509,67 +515,68 @@ List sampleCaNCPP(const int N,
 
 
 
+//' gmscale
+//' scaling of matrix with a geometric mean procedure
+//'MatrixXd A, VectorXd &cscale, VectorXd &rscale, double scltol
+//' @param A a matrix of coefficients of inequality constants A.x<=b
+//' @param b a vector of length equals to nrow(A)
+//' @param cscale the column scaling vector
+//' @param rscale the row scaling vector
+//' @param scltol tolerance
+//'
+//' @section Details:
+//' The CHRR algorithm is a C++ translation of cobratoolbox code written
+//' by Yin Zhang licensed under GNU GPL-3 https://github.com/opencobra/cobratoolbox/ 
+//' and of matlab code written by Ben Cousins 
+//' (https://github.com/Bounciness/Volume-and-Sampling) is a C++ translation of
+//' matlab functions provided in the opencobra toolbox and 
+//' in this github repository % `m x n` sparse matrix `A`.
+//' 
+//' An iterative procedure based on geometric means is used,
+//' following a routine written by Robert Fourer, 1979.
+//' Several passes are made through the columns and rows of `A`.
+//' The main steps are:
+//'
+//'   1. Compute :math:`aratio = max_j (max_i Aij / min_i Aij)`.
+//'   2. Divide each row `i` by :math:`sqrt( max_j Aij * min_j Aij)`.
+//'   3. Divide each column `j` by :math:`sqrt( max_i Aij * min_i Aij)`.
+//'   4. Compute `sratio` as in Step 1.
+//'   5. If :math:`sratio < aratio * scltol`,
+//'      set :math:`aratio = sratio` and repeat from Step 2.
+//'
+//' To dampen the effect of very small elements, on each pass,
+//' a new row or column scale will not be smaller than sqrt(damp)
+//' times the largest (scaled) element in that row or column.
+//'
+//' Use of the scales:
+//' To apply the scales to a linear program,
+//' :math:`min c^T x` st :math:`A x = b`, :math:`l \leq x \leq u`,
+//' we need to define "barred" quantities by the following relations:
+//' `A = R Abar C`, `b = R bbar`, `C cbar = c`,
+//' `C l = lbar`, `C u = ubar`, `C x = xbar`.
+//'
+//' This gives the scaled problem
+//' :math:`min\ cbar^T xbar` st :math:`Abar\ xbar = bbar`, :math:`lbar \leq xbar \leq ubar`.
+//'
+//' .. Author: - Michael Saunders, Systems Optimization Laboratory, Stanford University.
+//' ..
+//'    07 Jun 1996: First f77 version, based on MINOS 5.5 routine m2scal.
+//'    24 Apr 1998: Added final pass to make column norms = 1.
[…]
