+++ RCaNmodel/src/sampleCaNCPP.cpp
@@ -575,7 +575,10 @@ List sampleCaNCPP(const int N,
 
 
 
-void gmscale(MatrixXd A, VectorXd &cscale, VectorXd &rscale, double scltol){
+void gmscale(Eigen::MatrixXd A, 
+             Eigen::VectorXd &cscale, 
+             Eigen::VectorXd &rscale, 
+             double scltol){
 
 
   int m = A.rows();
@@ -587,24 +590,24 @@ void gmscale(MatrixXd A, VectorXd &cscale, VectorXd &rscale, double scltol){
   double damp=1e-4;
   double small = 1e-8;
   double eps=2.2204e-16;
-  MatrixXd SA;
-  rscale = VectorXd::Constant(m, 1);
-  cscale = VectorXd::Constant(n, 1);
+  Eigen::MatrixXd SA;
+  rscale = Eigen::VectorXd::Constant(m, 1);
+  cscale = Eigen::VectorXd::Constant(n, 1);
   
   /*--------------------------------------------------------------
    Main loop.
    --------------------------------------------------------------*/
-    MatrixXd invSA= MatrixXd::Zero(m,n);
+    Eigen::MatrixXd invSA= Eigen::MatrixXd::Zero(m,n);
   for (int npass = 0; npass <= maxpass; ++ npass){
     // Find the largest column ratio.
     // Also set new column scales (except on pass 0).
     
     rscale = (rscale.array() ==0).select(1, rscale);
-    MatrixXd Rinv    = rscale.cwiseInverse().asDiagonal();
+    Eigen::MatrixXd Rinv    = rscale.cwiseInverse().asDiagonal();
     SA      = Rinv*A;
     invSA= (SA.array()==0).select(1e9,SA.cwiseInverse());
-    VectorXd cmax    = SA.colwise().maxCoeff();   // column vector
-    VectorXd cmin    = invSA.colwise().maxCoeff();   // column vector
+    Eigen::VectorXd cmax    = SA.colwise().maxCoeff();   // column vector
+    Eigen::VectorXd cmin    = invSA.colwise().maxCoeff();   // column vector
     cmin    = (cmin.array() + eps).matrix().cwiseInverse();
     double sratio  = cmax.cwiseProduct(cmin.cwiseInverse()).maxCoeff();   // Max col ratio
     if (npass > 0){
@@ -623,8 +626,8 @@ void gmscale(MatrixXd A, VectorXd &cscale, VectorXd &rscale, double scltol){
     Eigen::DiagonalMatrix<double, Eigen::Dynamic> Cinv = cscale.cwiseInverse().asDiagonal();
     SA      = A*Cinv;                  // Scaled A
     invSA= (SA.array()==0).select(1e9,SA.cwiseInverse());
-    VectorXd rmax    = SA.rowwise().maxCoeff();   // column vector
-    VectorXd rmin    = invSA.rowwise().maxCoeff();   // column vector
+    Eigen::VectorXd rmax    = SA.rowwise().maxCoeff();   // column vector
+    Eigen::VectorXd rmin    = invSA.rowwise().maxCoeff();   // column vector
     rmin    = (rmin.array() + eps).matrix().cwiseInverse();
     rscale  = rmax.cwiseProduct(rmin.cwiseMax(damp*rmax)).array().sqrt().matrix();
   }
[…]
