+++ RCaNmodel/src/RcppExports.cpp
@@ -187,6 +187,201 @@ RcppExport SEXP _RCaNmodel_sampleCaNCPP(SEXP NSEXP, SEXP ASEXP, SEXP bSEXP, SEXP
     UNPROTECT(1);
     return rcpp_result_gen;
 }
+// gmscale
+void gmscale(Eigen::MatrixXd A, Eigen::VectorXd& cscale, Eigen::VectorXd& rscale, double scltol);
+static SEXP _RCaNmodel_gmscale_try(SEXP ASEXP, SEXP cscaleSEXP, SEXP rscaleSEXP, SEXP scltolSEXP) {
+BEGIN_RCPP
+    Rcpp::traits::input_parameter< Eigen::MatrixXd >::type A(ASEXP);
+    Rcpp::traits::input_parameter< Eigen::VectorXd& >::type cscale(cscaleSEXP);
+    Rcpp::traits::input_parameter< Eigen::VectorXd& >::type rscale(rscaleSEXP);
+    Rcpp::traits::input_parameter< double >::type scltol(scltolSEXP);
+    gmscale(A, cscale, rscale, scltol);
+    return R_NilValue;
+END_RCPP_RETURN_ERROR
+}
+RcppExport SEXP _RCaNmodel_gmscale(SEXP ASEXP, SEXP cscaleSEXP, SEXP rscaleSEXP, SEXP scltolSEXP) {
+    SEXP rcpp_result_gen;
+    {
+        Rcpp::RNGScope rcpp_rngScope_gen;
+        rcpp_result_gen = PROTECT(_RCaNmodel_gmscale_try(ASEXP, cscaleSEXP, rscaleSEXP, scltolSEXP));
+    }
+    Rboolean rcpp_isInterrupt_gen = Rf_inherits(rcpp_result_gen, "interrupted-error");
+    if (rcpp_isInterrupt_gen) {
+        UNPROTECT(1);
+        Rf_onintr();
+    }
+    bool rcpp_isLongjump_gen = Rcpp::internal::isLongjumpSentinel(rcpp_result_gen);
+    if (rcpp_isLongjump_gen) {
+        Rcpp::internal::resumeJump(rcpp_result_gen);
+    }
+    Rboolean rcpp_isError_gen = Rf_inherits(rcpp_result_gen, "try-error");
+    if (rcpp_isError_gen) {
+        SEXP rcpp_msgSEXP_gen = Rf_asChar(rcpp_result_gen);
+        UNPROTECT(1);
+        Rf_error(CHAR(rcpp_msgSEXP_gen));
+    }
+    UNPROTECT(1);
+    return rcpp_result_gen;
+}
+// mve_solver
+bool mve_solver(Eigen::MatrixXd A, Eigen::VectorXd b, const Eigen::VectorXd& x0, double reg, Eigen::VectorXd& x, Eigen::MatrixXd& E2, const int maxiter, const double tol);
+static SEXP _RCaNmodel_mve_solver_try(SEXP ASEXP, SEXP bSEXP, SEXP x0SEXP, SEXP regSEXP, SEXP xSEXP, SEXP E2SEXP, SEXP maxiterSEXP, SEXP tolSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::traits::input_parameter< Eigen::MatrixXd >::type A(ASEXP);
+    Rcpp::traits::input_parameter< Eigen::VectorXd >::type b(bSEXP);
+    Rcpp::traits::input_parameter< const Eigen::VectorXd& >::type x0(x0SEXP);
+    Rcpp::traits::input_parameter< double >::type reg(regSEXP);
[…]
