+++ RCaNmodel/NAMESPACE
@@ -26,8 +26,13 @@ export(ggSeries)
 export(ggTopDownBottomUp)
 export(ggTrophicRelation)
 export(ggViolin)
+export(gmscale)
+export(mve_run_cobra)
+export(mve_solver)
 export(plotPolytope2D)
+export(round)
 export(sampleCaN)
+export(shiftPolytope)
 export(toggleConstraint)
 export(writeCaN2XLS)
 importFrom(GGally,ggpairs)

[…]
