+++ README.md
@@ -2,7 +2,7 @@
 <img src="RCaNmodel/man/figures/logo.png" align="right" width="220" />
 
 ## RCaNmodel installation
-For RCaNmodel installation, please look [here](RCaN/README.md)
+For RCaNmodel installation, please look [here](RCaNmodel/README.md)
 
 ## RCaNconstructor installation
 For RCaNconstructor installation, please look [here](https://github.com/inrae/RCaNmodel/releases) (click on "Assets" in the section corresponding to your platform)
