### 2022-10-10 #4804 @fabian-s
Hi @eafulton,
Would you be willing to review this submission for JOSS?
