### 2022-11-02 #4804 @elfunesto
hi @fabian-s ,
Sorry for the delay. I had difficulty to find the github handles. I put the ones I found plus other potential reviewers
 K Soetaert =  [karlines](https://github.com/cran/limSolve/commits?author=karlines)
Pierre-Yves Hernvan (developer of R package ecoDiet) [pyhernvann](https://github.com/pyhernvann)
Matthew K Lau (developer of R package enaR)= [MKLau](https://github.com/SEELab/enaR/commits?author=MKLau)
Kelly Kearney (developer of a matlab version of ecopath)=[kakearney](https://github.com/kakearney/ecopathlite-pkg/commits?author=kakearney)
