### 2022-11-07 #4804 @fabian-s
hi @sgaichas @slucey  @andybeet

Would you be willing to review this submission for JOSS?
