### 2022-11-08 #4804 @karlines
Hi, this is not my expertise, so I will decline.

From: Fabian Scheipl ***@***.***>
Sent: Monday, November 7, 2022 10:48 AM
Cc: Karline Soetaert ***@***.***>; Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [PRE REVIEW]: RCaNmodel: a R package for Chance and Necessity modelling (Issue #4804)

hi @sgaichas<https://eur01.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2Fsgaichas&data=05%7C01%7C%7Cb6961ec4d1c84e902b6308dac0a52307%7C9a1651bf58af435b86a83e9334b4b732%7C0%7C0%7C638034112732506860%7CUnknown%7CTWFpbGZsb3d8eyJWIjoiMC4wLjAwMDAiLCJQIjoiV2luMzIiLCJBTiI6Ik1haWwiLCJXVCI6Mn0%3D%7C3000%7C%7C%7C&sdata=vwQ9LN8lwhfGm12bweu3nLYD0yBOnm7EoO03qq1l7yU%3D&reserved=0> @slucey<https://eur01.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2Fslucey&data=05%7C01%7C%7Cb6961ec4d1c84e902b6308dac0a52307%7C9a1651bf58af435b86a83e9334b4b732%7C0%7C0%7C638034112732506860%7CUnknown%7CTWFpbGZsb3d8eyJWIjoiMC4wLjAwMDAiLCJQIjoiV2luMzIiLCJBTiI6Ik1haWwiLCJXVCI6Mn0%3D%7C3000%7C%7C%7C&sdata=Bs7TKX9LPHnZnmwl6NOx%2FfZnXEJK%2BZeUnCas3%2BaSMwo%3D&reserved=0> @andybeet<https://eur01.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2Fandybeet&data=05%7C01%7C%7Cb6961ec4d1c84e902b6308dac0a52307%7C9a1651bf58af435b86a83e9334b4b732%7C0%7C0%7C638034112732506860%7CUnknown%7CTWFpbGZsb3d8eyJWIjoiMC4wLjAwMDAiLCJQIjoiV2luMzIiLCJBTiI6Ik1haWwiLCJXVCI6Mn0%3D%7C3000%7C%7C%7C&sdata=O30jlLKzUA1ivP62SRXkL9uiKIrNZaROC4NDpvxBDl0%3D&reserved=0>

Would you be willing to review this submission for JOSS?
We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html<https://eur01.safelinks.protection.outlook.com/?url=https%3A%2F%2Fjoss.readthedocs.io%2Fen%2Flatest%2Freview_criteria.html&data=05%7C01%7C%7Cb6961ec4d1c84e902b6308dac0a52307%7C9a1651bf58af435b86a83e9334b4b732%7C0%7C0%7C638034112732663096%7CUnknown%7CTWFpbGZsb3d8eyJWIjoiMC4wLjAwMDAiLCJQIjoiV2luMzIiLCJBTiI6Ik1haWwiLCJXVCI6Mn0%3D%7C3000%7C%7C%7C&sdata=nKxlwV2VMFUB9NNRdZkuM%2BSoyNfYxNav9YQxKaPOQ88%3D&reserved=0>

-
