### 2022-11-08 #4804 @pyhernvann
Dear Fabian,
Sorry for the late answer - Unfortunately I am currently reviewing several
studies and have to decline unless it can wait until mid-December.
Best regards,
PYH

Le mer. 2 nov. 2022 à 07:18, Fabian Scheipl ***@***.***> a
écrit :

> Hi @karlines <https://github.com/karlines> @MKLau
> <https://github.com/MKLau> @pyhernvann <https://github.com/pyhernvann>
> Would you be willing to review this submission for JOSS?
> We carry out our checklist-driven reviews here in GitHub issues and follow
> these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
