### 2022-11-08 #4804 @fabian-s
Hi @amoeba,
hi @pboesu,
hi @jhollist :wave:

Would you be willing to review [this submission](https://raw.githubusercontent.com/openjournals/joss-papers/joss.04804/joss.04804/10.21105.joss.04804.pdf) for JOSS?
