### 2022-11-14 #4804 @fabian-s
hi @jhollist @trashbirdecology @SteveViss wave

Would you be willing to review [this submission](https://raw.githubusercontent.com/openjournals/joss-papers/joss.04804/joss.04804/10.21105.joss.04804.pdf) for JOSS?
