### 2022-11-15 #4804 @Kevin-Mattheus-Moerman
> @Kevin-Mattheus-Moerman as you can see I'm having a _very_ hard time finding 2 reviewers for this. how about I do the 2nd review myself to keep this moving?

As a last resort I would say that is okay. We can also perhaps stress to reviewers that we can be flexible regarding time needed. Some responses you got :point_up: seem to say "I do not have time now/this month", perhaps we can say, how about next month? I know it delays things but it is better than no reviewers. But again, if you can be the second reviewer then that would be permitted as a last resort. I have had to do that in the past on 2 occasions.
