### 2022-11-15 #4804 @trashbirdecology
@gabspalomo -- consider becoming a JOSS reviewer?

*______________________________________________________*

Jessica L. Burnett (she/her)

Github <http://github.com/trashbirdecology>

ORCID: 0000-0002-0896-5099
<https://orcid.org/0000-0002-0896-5099>
*______________________________________________________*

Join the Community for Data Integration <https://www.usgs.gov/centers/cdi>! All
welcome. It's cool.

On Tue, Nov 15, 2022 at 7:57 AM Fabian Scheipl ***@***.***>
wrote:

> @Kevin-Mattheus-Moerman <https://github.com/Kevin-Mattheus-Moerman> as
> you can see I'm having a *very* hard time finding 2 reviewers for this.
> how about I do the 2nd review myself to keep this moving?
>
