### 2022-11-22 #4955 @editorialbot
[REVIEW]: RCaNmodel: a R package for Chance and Necessity modelling
**Submitting author:** @elfunesto (<a href="http://orcid.org/0000-0001-9206-0040">Hilaire Drouineau</a>)
**Repository:** https://github.com/inrae/RCaNmodel/
**Branch with paper.md** (empty if default branch): master
**Version:** v2.0.2
