### 2022-11-22 #4955 @fabian-s
@elfunesto please check and fix the missing/invalid DOIs found by the bot
