+++ paper/paper.bib
@@ -2,6 +2,7 @@
 @article{pauly2000,
 	title = {Ecopath, {Ecosim}, and {Ecospace} as tools for evaluating ecosystem impact of fisheries},
 	volume = {57},
+	doi = {10.1006/jmsc.2000.0726},
 	url = {http://icesjms.oxfordjournals.org/content/57/3/697.short},
 	number = {3},
 	urldate = {2015-11-18},
@@ -31,6 +32,7 @@ type: Journal article},
 	volume = {42},
 	number = {3},
 	journal = {Marine ecology progress series. Oldendorf},
+	doi = {10.3354/MEPS042269},
 	author = {Vézina, Alain F. and Platt, Trevor},
 	year = {1988},
 	pages = {269--287},
@@ -41,6 +43,7 @@ type: Journal article},
 	title = {Gaining insight into food webs reconstructed by the inverse method},
 	volume = {60},
 	number = {1-2},
+  doi = {10.1016/j.jmarsys.2005.12.002},
 	journal = {Journal of Marine Systems},
 	author = {Kones, Julius K. and Soetaert, Karline and van Oevelen, Dick and Owino, John O. and Mavuti, Kenneth},
 	year = {2006},
@@ -52,6 +55,7 @@ type: Journal article},
 @article{kones2009,
 	title = {Are network indices robust indicators of food web functioning? {A} {Monte} {Carlo} approach},
 	volume = {220},
+	doi = {10.1016/j.ecolmodel.2008.10.012},
 	shorttitle = {Are network indices robust indicators of food web functioning?},
 	number = {3},
 	journal = {Ecological Modelling},
@@ -117,6 +121,7 @@ tex.ids= addison2013a},
 @incollection{niquil2011,
 	address = {Waltham, MA},
 	title = {Inverse modeling in modern ecology and application to coastal ecosystems},
+	doi = {10.1016/B978-0-12-374711-2.00906-2},
 	booktitle = {Treatise on estuarine and coastal science.},
 	publisher = {Academic Press},
 	author = {Niquil, N. and Saint-Béat, B. and Johnson, G. A. and Soetaert, K. and Van Oevelen, D. and Bacher, C. and Vézina, A. F.},
@@ -220,7 +225,7 @@ tex.ids= addison2013a},
 	title = {Best practice in {Ecopath} with {Ecosim} food-web models for ecosystem-based management},
 	volume = {331},
 	issn = {0304-3800},
-	doi = {http://dx.doi.org/10.1016/j.ecolmodel.2015.12.007},
+	doi = {10.1016/j.ecolmodel.2015.12.007},
 	journal = {Ecological Modelling},
 	author = {Heymans, Johanna Jacomina and Coll, Marta and Link, Jason S. and Mackinson, Steven and Steenbeek, Jeroen and Walters, Carl and Christensen, Villy},
 	year = {2016},
