### 2022-11-22 #4955 @elfunesto
@fabian-s This is addressed with commit [inrae/RCaNmodel.git@5cb4842](https://github.com/inrae/RCaNmodel/commit/1ff07450912bf7d2314783e80df9d10d58a7b1e7)
