### 2023-01-18 2250501
Prefer remotes over devtools for install_github
remotes is smaller and devtools actually uses remotes under the hood. devtools is a much larger package than remotes so this approach should be faster/easier for more users.
