### 2023-01-18 bde97e8
Copy edit RCaNconstructor README
