### 2023-01-18 inrae/RCaNmodel#2 @amoeba
JOSS Review
Hello @elfunesto, I've had a chance to review this work for https://github.com/openjournals/joss-reviews/issues/4955 and would like to request some changes.

I was able to check off everything except the following. Please address them:

- [X] License: Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license?
  - [X] Your LICENSE files contain some boilerplate that should be removed (lines 1-16 in each)
  - [X] I don't think this submission currently meets the standard outline in the [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#documentation). I suggest the authors re-work the three READMEs (top level, one in RCaNModel, and one in RCaNconstructor) so the top level README comprehensively covers the criteria (Statement of need, installation instructions, usage, community guidelines). Links between Markdown docs at the top level and in subfolders would be fine but I'd like to see the top level README be fairly comprehensive.
  - [X] I found this in the RCaNmodel README but not in the top level README which is a more common place to see it. Please include this in the top level README re-work requested above.

As I reviewed, I found other places to provide my feedback which aren't captured by the [JOSS Review Criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#documentation). You can consider these suggestions:

- [X] License
  - [X] I suggest you add the `License` field to your `DESCRIPTION file as outlined in https://r-pkgs.org/license.html
- [X] Contribution and authorship
  - [X] While poking around the repo, I noticed your commit log is a bit messy with respect to authors. I suggest coordinating across the group to make sure everyone's git configs are correct. You might also consider rewriting history to clean things up.
    ```
    $ git shortlog -sn --all -e
	     510  hilaire drouineau <hilaire drouineau@irstea.fr>
	      44  elfunesto <43066644+elfunesto@users.noreply.github.com>
	      37  Christian Mullon <ch.mullon@gmail.com>
	      10  Planque, Benjamin <benjamin.planque@hi.no>
	       3  christianmullon <ch.mullon@gmail.com>
	       2  Benjamin Planque <benjamin.planque@hi.no>
	       2  Drouineau Hilaire <hilaire.drouineau@irstea.fr>
	       2  mullon-ch <64130573+mullon-ch@users.noreply.github.com>
	       1  Mullon <ch.mullon@gmail.com>
	       1  Planque <benjamin.planque@hi.no>
    ```
- [X] Automated tests
  - [X] I noticed one of your tests is failing `Failure (test-checkPolytopeStatus.R:10): this polytope is unbounded`
- [X] Your `DESCRIPTION` file has default text in the Description field
    https://github.com/inrae/RCaNmodel/blob/1ff07450912bf7d2314783e80df9d10d58a7b1e7/RCaNmodel/DESCRIPTION#L10
- [X] The repo is quite large. I wasn't able to clone the repository on my home internet connection had to go find a faster one. I noticed you've committed a lot of binary assets to the repo which is likely the cause of the size (>1GB). Unless you have a specific reason to do this, I suggest cleaning the repo up (i.e., rewriting history) and storing assets in more appropriate places such as GitHub Releases or software archives.
- [X] The vignette covered what I think is a reasonable example of using your software but I think it could use quite a bit of attention to make it more readable. Some of this is your choice of API design (return values, offering or not offering print() methods) and some is how the RMarkdown doc is constructed (you could use tables more, where appropriate).
