### 2023-01-18 #4955 @amoeba
Hello @fabian-s and @elfunesto, I've conducted my review and I am requesting changes from @elfunesto over on [the RCaNmodel repo](https://github.com/inrae/RCaNmodel/issues/2). At this point, my review status is "Minor Revisions".
