### 2023-01-18 ea852bc
Update README.md
many users already have devtools installed, so we propose both devtools and remote
