### 2023-01-18 0723dca
remove boileplates in LICENSE files Ref #2
