### 2023-01-18 97270c3
add a statement of need in the main readme, community guideline and usage instructions in the main readme Ref #2
