+++ README.md
@@ -1,5 +1,16 @@
-# CaN
+# CaNmodel
 <img src="RCaNmodel/man/figures/logo.png" align="right" width="220" />
+
+## What is CaNmodel aimed at?
+RCaNmodel and RCaNmodelconstructor primarily aim at implenting food-web models based on a
+Chance and Necessity approach. Details and relation with pre-existing tools can be found in 
+the statement of need of this [document](paper/paper.md).
+As a summary, those model focus on trophic exchanges between preys and predators while accouting
+for the intrinsic stochasticity of nature and the existence of physical laws. The R package RCaNmodel 
+facilitate the modeling process (model construction and validation, model exploration, output analysis),
+while the RCaNconstructor GUI provides an interactive/intuitive interface to support 
+a participatory framework.
+ 
 
 ## RCaNmodel installation
 For RCaNmodel installation, please look [here](RCaNmodel/README.md)
