### 2023-01-18 inrae/RCaNmodel#2 @elfunesto
Dear @amoeba , Thank you for your feedback. I am trying to revise the package following your recommendations. Small question regarding the license field in the description file. I have already such a field (line 11)  in RCaNmodel/DESCRIPTION, isn'it enought?
