### 2023-01-18 f1a2f0e
fix the description field in RCaNmodel/DESCRIPTION Ref #2
