### 2023-01-18 #4955 @fabian-s
hi @juanvillada, please update us on how your review is going.

if boxes in your checklist remain unchecked because of problems you have identified, please open respective issues in the [RCaNmodel repo](https://github.com/inrae/RCaNmodel/issues) with some explanations so @elfunesto can fix these -- JOSS aims for a collaborative and constructive review experience.
