### 2023-01-18 inrae/RCaNmodel#2 @elfunesto
I have used filter repo to remove unused filed, tidy the commit history, and I have used lfs migrate to move binary files to lfs
