### 2023-01-18 inrae/RCaNmodel#2 @elfunesto
@amoeba Two additional questions:

- the test successfully passes here. It is very weird since it is a very simple test using lpsolve facility. Can you tell please me the output of `checkPolytopeStatus(list(A = diag(4)[-4, ],
                                         b = rep(1, 3)))` ?
- regarding the vignette, we have added an additional Rmarkdown alongside the paper which is I think of better quality than the former. Would it be ok if that Rmardown becomes the new vignette? It is located in paper/barents_SM_RMD.Rmd
