+++ .github/workflows/maven-package.yml
@@ -16,6 +16,8 @@ jobs:
 
     steps:
     - uses: actions/checkout@v2
+      with:
+        lfs: 'true'
     - name: Set up JDK 17
       uses: actions/setup-java@v2
       with:
