+++ .github/workflows/maven-build-installer-unix.yml
@@ -15,6 +15,8 @@ jobs:
 
     steps:
     - uses: actions/checkout@v2
+      with:
+        lfs: 'true'
     - name: Set up JDK 17
       uses: actions/setup-java@v2
       with:
