+++ .github/workflows/maven-build-installer-windows.yml
@@ -28,6 +28,8 @@ jobs:
     - name: Add Wix to Path
       run: echo "$HOME/target/wix" >> $GITHUB_PATH
     - uses: actions/checkout@v2
+      with:
+        lfs: 'true'
     - name: Set up JDK 17
       uses: actions/setup-java@v2
       with:
