### 2023-01-18 inrae/RCaNmodel#2 @amoeba
> I have already such a field (line 11) in RCaNmodel/DESCRIPTION, isn'it enought?

Yes, I must have missed it. This looks good.

> I have used filter repo to remove unused filed, tidy the commit history, and I have used lfs migrate to move binary files to lfs

Thanks. It's a bit better.

> the test successfully passes here. It is very weird since it is a very simple test using lpsolve facility. Can you tell please me the output of checkPolytopeStatus(list(A = diag(4)[-4, ], b = rep(1, 3))) ?

I get:

```
> checkPolytopeStatus(list(A = diag(4)[-4, ], b = rep(1, 3)))
[1] "polytope ok"
```

> regarding the vignette, we have added an additional Rmarkdown alongside the paper which is I think of better quality than the former. Would it be ok if that Rmardown becomes the new vignette? It is located in paper/barents_SM_RMD.Rmd

Maybe this one would be good as a new vignette instead of replacing the current main vignette since it's fairly long.
