### 2023-01-20 inrae/RCaNmodel#2 @amoeba
I'm on macOS 13.1, arm64 arch. You might consider adding https://github.com/r-lib/actions/blob/v2/examples/check-standard.yaml to your GitHub Actions workflows so check+test run on different platforms and versions.
