+++ RCaNmodel/vignettes/vignette_RCaN.Rmd
@@ -1,6 +1,7 @@
 ---
-title: "Tutorial of RCaNmodel"
-output:
+title: 'An application of CaN modelling to the Barents  Sea trophic network'
+date: "5/2/2021"
+output: 
   rmarkdown::html_vignette:
      keep_md: yes
   pdf_document:
@@ -10,184 +11,493 @@ author:
 - "Benjamin Planque - IMR"
 - "Christian Mullon - IRD"
 vignette: >
-  %\VignetteIndexEntry{Tutorial of RCaNmodel}
+  %\VignetteIndexEntry{An application of CaN modelling to the Barents  Sea trophic network}
   %\VignetteEncoding{UTF-8}
   %\VignetteEngine{knitr::rmarkdown}
+fig_width: 8
+bibliography: CaN.bib
 ---
 
-```{r, include = FALSE}
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>"
-)
-```
+```{r label=setup, include=FALSE}
 
-```{r setup}
-library(RCaNmodel)
-```
-
-This package aims at fitting a chance and necessity ecosystem model. A full description of this kind of model is provided in 
-
-Planque, B., & Mullon, C. (n.d.). Modelling chance and necessity in natural systems. ICES Journal of Marine Science. https://doi.org/10.1093/icesjms/fsz173
-
-
-# Data requirements
-Inputs should be provided in an xlxs data file. This file can be generated using an independent java graphical user interface, but it can be created independently. An artificial file is provided in the package as a template and as an example for this vignette. You can find the path to the file by entering
-
-```{r}
-system.file("extdata", "CaN_template_mini.xlsx", package = "RCaNmodel")
-```
-
-```{r,echo=FALSE,include=FALSE}
-myCaNmod <- buildCaN(system.file("extdata", "CaN_template_mini.xlsx", package = "RCaNmodel"))
-```
-
-There are a few important things to remember:
-* the names of the sheets should match the names of the sheets in the template (i.e. "Components & input parameter", "Fluxes", "Input time-series", "Constraints")
-* the names of the columns in sheets "Components & input parameter", "Fluxes" and "Constraints" and there should be a column "Year" in the sheet "Input time-series"
-* it is worthwile giving as much information as possible in the "Info" sheet
-* you can add other sheets and other columns in your file is needed, as the minimal sets of sheets and columns are provided
-
-Note that the object can also build directly from R using the buildCaN function.
-
-## Sheet "Components & input parameter"
-```{r,echo=FALSE,include=TRUE}
-myCaNmod$components_param
-```
-Components correspond to the boxes of the trophic food web. Most components are internal to the ecosystem (denoted 1), but other can be external, for example to deal with fishery or migrations, and are denoted 0. The definition of the parameters are provided in Planque and Mullon (in press). You can leave some cells empty:
-* if you do not want to use a specific constraint (for example, if you do not want to use a satiety constraint for a component, leave the cell empty)
-* for external component, you can leave all cells empty since the dynamics of external components are not modelled, except digestibility of components that are eated by internal components (for example in the template, we have a digestibility for PhytoAndBacteria).
-
-## Sheet "Fluxes"
-```{r,echo=FALSE,include=TRUE}
-myCaNmod$fluxes_def
-```
-Column "Flux" aims at providing name to identify the differents fluxes. Then, "From" and "To" define the different (oriented) flow in the food web. The names in "From" and "To" should match the names in Components (case sensitive). Trophic is a boolean that indicates whether the flow corresponds to a trophic flow or to a non trophic flow (e.g. migration, fishery etc.). Contrary to non trophic flows, a trophic flow will be submitted to digestibility and assimilation.
-
-
-## Sheet "Input time-series"
[…]
