+++ RCaNmodel/DESCRIPTION
@@ -32,6 +32,7 @@ Imports:
     RcppEigen
 LinkingTo: Rcpp, RcppEigen, dqrng
 Suggests: 
+    kableExtra,
     gridExtra,
     knitr,
     rmarkdown,
