+++ .github/workflows/check-standard.yaml
@@ -50,7 +50,10 @@ jobs:
         with:
           extra-packages: any::rcmdcheck
           needs: check
-
+          working-directory: RCaNmodel 
+          
       - uses: r-lib/actions/check-r-package@v2
         with:
+          args: 'c("--no-manual", "--as-cran")'
           upload-snapshots: true
+          working-directory: RCaNmodel
