+++ .github/workflows/check-standard.yaml
@@ -48,7 +48,7 @@ jobs:
 
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
-          extra-packages: any::rcmdcheck
+          extra-packages: any::rcmdcheck, ROI.plugin.cbc=?ignore
           needs: check
           working-directory: RCaNmodel 
           
