### 2023-01-23 inrae/RCaNmodel#2 @elfunesto
Thanks a lot. I set up the github action
