+++ .github/workflows/maven-build-installer-unix.yml
@@ -14,7 +14,7 @@ jobs:
     runs-on: ubuntu-latest
 
     steps:
-    - uses: actions/checkout@v2
+    - uses: actions/checkout@v3
       with:
         lfs: 'true'
     - name: Set up JDK 17
