+++ .github/workflows/maven-build-installer-windows.yml
@@ -27,7 +27,7 @@ jobs:
         pathTarget: ./target/wix
     - name: Add Wix to Path
       run: echo "$HOME/target/wix" >> $GITHUB_PATH
-    - uses: actions/checkout@v2
+    - uses: actions/checkout@v3
       with:
         lfs: 'true'
     - name: Set up JDK 17
