### 2023-01-24 #4955 @juanvillada
Hi @fabian-s and @elfunesto,

Thanks for inviting me to review this work. As I mentioned before, my revision is merely technical and not conceptual as I am not an expert in this field. The manuscript thoroughly presents the RCaNmodel package, it provides examples of usage, and yields good visualization for the diagnostics diagrams. Overall, I think it is a good technical contribution to the Open Source Software community.

Just some minor suggestions here:

- [ ] ⚠️ **License**. Please include the LICENSE in the main repository. I found a LICENSE file but it is placed in https://github.com/inrae/RCaNmodel/blob/master/RCaNmodel/LICENSE and not in the main repository folder.

- [ ] ⚠️ **A statement of need**.. This statement is clear from the manuscript. I would like to suggest to include a short statement of need in the repository README to catch better the attention of users.

- [ ] ⚠️ **Community guidelines**.. Please reframe the "Bug reporting" section (https://github.com/inrae/RCaNmodel#bug-reporting) to better convey:
>Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

- [ ] ⚠️ **References**.. There are 3 references missing the DOI. I assume they don't exist. Is that correct?

Thanks!
