### 2023-01-25 f0f0e59
add two missing DOI
