### 2023-01-25 #4955 @elfunesto
Hi @juanvillada, Thank you for your review. Here are a few replies:

- References: I added few missing DOI in commit #f0f0e59. The remaining references that do not have DOI are manuals or techreport for which DOI do not exist to my knowledge.
- Statement of need, following @amoeba suggestion, I added a statement of need in the main README. It is already [online](https://github.com/inrae/RCaNmodel) . Is this ok for you?
- Community guidelines: I added and enhanced a community guidelines in the main readme (same page) in commit f8f4a2c. Is this ok?
- License: the repository has two pieces of software (the R package and the java GUI) which have their own licenses (BSD3, same copyright) in their own subdirectories. Should I add a license file that refer links to the two licenses file?
