### 2023-01-25 inrae/RCaNmodel#2 @elfunesto
Hi @amoeba ,
Funny discovery: the test was working in most machines, including mac machines from r-hub. One of the co-author as the same machine as you so we were to try... and found what is likely to be a R bug on your architecture! In my code, I have a test with 1e30 which corresponds to infinite value in lpsolve. But looks at the difference between 1.0e30 and 1e30! it is not an epsilon and the problem does not happen with 1e29. Surprising! I implement a workaround in the package.

`Browse[3]> any(x0 == 1.0e30)
[1] FALSE
Browse[3]> x0
[1] 1e+00 1e+00 1e+00 1e+30
Browse[3]> any(x0 >= 1.0e30)
[1] TRUE
Browse[3]> 1e30==1.0e30
[1] FALSE
Browse[3]> 1e30-1.0e30
[1] 1.407375e+14
Browse[3]> 1e31-1.0e31
[1] 0
Browse[3]> 1e29*10-1.0e30
[1] 0`
