+++ .github/workflows/maven-build-installer-macos.yml
@@ -4,10 +4,12 @@
 name: Build macOS Installer
 
 on:
+  workflow_dispatch:
+    branches: [ master ]
   push:
     branches: [ master ]
-  pull_request:
-    branches: [ master ]
+  paths:
+    - RCaNconstructor/**
 
 jobs:
   build:
