+++ .github/workflows/maven-build-installer-unix.yml
@@ -4,6 +4,8 @@
 name: Build Unix Installer
 
 on:
+  workflow_dispatch:
+    branches: [ master ]
   push:
     branches: [ master ]
   pull_request:
