+++ .github/workflows/maven-build-installer-windows.yml
@@ -29,9 +29,8 @@ jobs:
         pathTarget: ./target/wix
     - name: Add Wix to Path
       run: echo "$HOME/target/wix" >> $GITHUB_PATH
-    - uses: actions/checkout@v3
-      with:
-        lfs: 'true'
+    - name: Checkout code
+      uses: nschloe/action-cached-lfs-checkout@v1
     - name: Set up JDK 17
       uses: actions/setup-java@v2
       with:
