+++ .github/workflows/maven-build-installer-macos.yml
@@ -17,9 +17,8 @@ jobs:
     runs-on: macos-latest
 
     steps:
-      - uses: actions/checkout@v3
-        with:
-          lfs: 'true'
+      - name: Checkout code
+        uses: nschloe/action-cached-lfs-checkout@v1
       - name: Set up JDK 17
         uses: actions/setup-java@v2
         with:
