+++ .github/workflows/maven-build-installer-unix.yml
@@ -16,9 +16,8 @@ jobs:
     runs-on: ubuntu-latest
 
     steps:
-    - uses: actions/checkout@v3
-      with:
-        lfs: 'true'
+    - name: Checkout code
+      uses: nschloe/action-cached-lfs-checkout@v1
     - name: Set up JDK 17
       uses: actions/setup-java@v2
       with:
