+++ .github/workflows/maven-package.yml
@@ -6,8 +6,9 @@ name: Maven Basic Package
 on:
   push:
     branches: [ master ]
-  pull_request:
-    branches: [ master ]
+  paths:
+    - RCaNconstructor/**
+    
 
 jobs:
   build:
@@ -15,9 +16,8 @@ jobs:
     runs-on: macos-latest
 
     steps:
-    - uses: actions/checkout@v3
-      with:
-        lfs: 'true'
+    - name: Checkout code
+      uses: nschloe/action-cached-lfs-checkout@v1
     - name: Set up JDK 17
       uses: actions/setup-java@v2
       with:
