+++ .github/workflows/check-standard.yaml
@@ -35,9 +35,8 @@ jobs:
       R_KEEP_PKG_SOURCE: yes
 
     steps:
-      - uses: actions/checkout@v3
-        with:
-          lfs: 'true'
+      - name: Checkout code
+        uses: nschloe/action-cached-lfs-checkout@v1
 
       - uses: r-lib/actions/setup-pandoc@v2
 
