+++ .github/workflows/maven-package.yml
@@ -17,7 +17,10 @@ jobs:
 
     steps:
     - name: Checkout code
-      uses: nschloe/action-cached-lfs-checkout@v1
+      uses: rohmanngmbh/action-checkout-repo@v1.3.2
+      with:
+        lfs: true
+        path: RCaNconstructor
     - name: Set up JDK 17
       uses: actions/setup-java@v2
       with:
