+++ .github/workflows/maven-build-installer-windows.yml
@@ -30,7 +30,10 @@ jobs:
     - name: Add Wix to Path
       run: echo "$HOME/target/wix" >> $GITHUB_PATH
     - name: Checkout code
-      uses: nschloe/action-cached-lfs-checkout@v1
+      uses: rohmanngmbh/action-checkout-repo@v1.3.2
+      with:
+        lfs: true
+        path: RCaNconstructor
     - name: Set up JDK 17
       uses: actions/setup-java@v2
       with:
