+++ .github/workflows/check-standard.yaml
@@ -36,7 +36,10 @@ jobs:
 
     steps:
       - name: Checkout code
-        uses: nschloe/action-cached-lfs-checkout@v1
+        uses: rohmanngmbh/action-checkout-repo@v1.3.2
+        with:
+          lfs: true
+          path: RCaNmodel
 
       - uses: r-lib/actions/setup-pandoc@v2
 
