### 2023-01-27 inrae/RCaNmodel#2 @amoeba
Very interesting. I would have though it was an epsilon issue like you mentioned but the difference is very large. Even code like this (on arm64) is strange to me:

```r
> format(1e22, scientific=FALSE)
[1] "10000000000000000000000"
> format(1e23, scientific=FALSE)
[1] " 99999999999999991611392"
```

It looks like you've gotten all of my checklist items checked off. I'll make another review pass before this Monday and update my review status.
