+++ paper/paper.md
@@ -2,7 +2,7 @@
 title: 'RCaNmodel: An R package for Chance and Necessity modelling'
 tags:
   - R
-  - tophic food web model
+  - trophic food web model
   - javafx
   - trophic controls
   - linear inverse modelling
@@ -33,7 +33,7 @@ bibliography: paper.bib
 
 RCaNmodel is an R package dedicated to the development and exploration of time-dynamic food-web models.
 The associated user-friendly JavaFX graphical user interface (GUI) RCaNconstructor is specifically developed to facilitate the
-co-construction of such models with non-modellers. Together, RCaNmodel and RCaNconstructor ensure the
+co-construction of such models with non-modelers. Together, RCaNmodel and RCaNconstructor ensure the
 traceability and replicability of food-web model developments, and provide a fast and efficient solution
 to handle the complex mathematical steps that arise from dynamic linear inverse modelling and polytope sampling
 in high dimensions. The software includes standard graphical outputs for the exploration of food-webs
@@ -50,7 +50,7 @@ have no single best solutions, but rather an infinite number of flow patterns th
 data, and food-web models are said to be underdetermined [@niquil2011]. Several strategies have been
 developed to tackle this issue. In Ecopath [@Polovina1984], or in its dynamic extension called Ecosim [@pauly2000; @christensen2004]
 one of the most largely used trophic food-web
-models [@Heymans2016] the modeller is asked to provide additional information on species diet or on their
+models [@Heymans2016] the modeler is asked to provide additional information on species diet or on their
 vulnerabilities to turn the underdetermined into an overdetermined problem. This approach ignores the highly
 variable and stochastic nature of trophic interactions. Food-web models based on linear inverse modelling (LIM) have been
 proposed as an alternative approach [@vezina1988]. LIM models rely on the same key assumptions as of any
@@ -126,6 +126,6 @@ possible time-trajectories of the food-web and that can be sampled using appropr
 be achieved using either an R script or the RCaNconstructor GUI, except the step denoted with a \*. Italicised texts denote
 *RCaNmodel* functions. The boxes stand for the main steps, under which is mentioned their links with other R libraries.\label{fig:approach}](approaches.png)
 
-![Example of diagnostic diagrams provided by *RCaNmodel*, based on the exemple provided as a supplementary material. Panel a shows the time variations of the biomass of a tophic group (DemF =  demersal fishes) and the trophic flow from PelF (Pelagic fishes) to DemF (the lines correspond to 3 random trajectories while the ribbons corresponds to quantiles 50%, 97.5% and 100% among trajectories). Panel b illustrates how the growth rate of DemF (y-axis) decreases with the biomass of the stock (x-axis) among the different years and MCMC iterations, suggesting a limitation by resources. Panel c illustrates that the trophic flow towards DemF is correlated to the amount of available preys, suggesting a bottom-up control of DemF. Panel d shows that the growth of DemF (MM) is positively correlated to its feeding, suggesting a strong bottom-up control. On the other hand, the growth rate is not correlated to the predation by predators suggesting that there is no top-down control.\label{fig:outupfig}](figure3.png)
[…]
