### 2023-01-31 #4955 @fabian-s
thanks @amoeba @elfunesto for your constructive and dilligent reviews :+1:

LICENSE files seem fine to me as is.

Looks like this is good to go once my PR https://github.com/inrae/RCaNmodel/pull/3 with some typo fixes has been integrated.
