+++ .github/workflows/maven-build-installer-unix.yml
@@ -8,8 +8,8 @@ on:
     branches: [ master ]
   push:
     branches: [ master ]
-  pull_request:
-    branches: [ master ]
+  paths:
+    - RCaNconstructor/**
 
 jobs:
   build:
