### 2023-02-01 b0d7396
add CITATION.cff
