+++ .github/workflows/maven-build-installer-macos.yml
@@ -4,8 +4,6 @@
 name: Build macOS Installer
 
 on:
-  workflow_dispatch:
-    branches: [ master ]
   push:
     branches: [ master ]
   paths:
