+++ .github/workflows/maven-package.yml
@@ -4,8 +4,6 @@
 name: Maven Basic Package
 
 on:
-  workflow_dispatch:
-    branches: [ master ]
   push:
     branches: [ master ]
   paths:
