+++ .github/workflows/maven-build-installer-unix.yml
@@ -6,8 +6,8 @@ name: Build Unix Installer
 on:
   push:
     branches: [ master ]
-  paths:
-    - RCaNconstructor/**
+    paths:
+      - RCaNconstructor/**
 
 jobs:
   build:
