+++ .github/workflows/maven-build-installer-macos.yml
@@ -6,8 +6,8 @@ name: Build macOS Installer
 on:
   push:
     branches: [ master ]
-  paths:
-    - RCaNconstructor/**
+    paths:
+      - RCaNconstructor/**
 
 jobs:
   build:
@@ -15,11 +15,7 @@ jobs:
     runs-on: macos-latest
 
     steps:
-      - name: Checkout code
-        uses: rohmanngmbh/action-checkout-repo@v1.3.2
-        with:
-          lfs: true
-          path: RCaNconstructor
+      - uses: actions/checkout@v2
       - name: Set up JDK 17
         uses: actions/setup-java@v2
         with:
