+++ .github/workflows/maven-build-installer-windows.yml
@@ -4,6 +4,8 @@
 name: Build Windows Installer
 
 on:
+  workflow_dispatch:
+    branches: [ master ]
   push:
     branches: [ master ]
     paths:
