+++ CITATION.cff
@@ -1,15 +1,0 @@
-cff-version: 1.1.0
-message: "If you use this software, please cite it as below."
-authors:
-  - family-names: Drouineau
-    given-names: Hilaire
-    orcid: https://orcid.org/0000-0001-9206-0040
-  - family-names: Planque
-    given-names: Benjamin
-    orcid: https://orcid.org/0000-0002-0557-7410
-  - family-names: Mullon
-    given-names: Christian
-    orcid: https://orcid.org/0000-0003-3990-9118
-title: inrae/RCaNmodel: RCaNmodel: An R package for Chance and Necessity modelling
-version: v2.0
-date-released: 2023-02-01
