+++ .github/workflows/maven-build-installer-unix.yml
@@ -16,11 +16,7 @@ jobs:
     runs-on: ubuntu-latest
 
     steps:
-    - name: Checkout code
-      uses: rohmanngmbh/action-checkout-repo@v1.3.2
-      with:
-        lfs: true
-        path: RCaNconstructor
+    - uses: actions/checkout@v3
     - name: Set up JDK 17
       uses: actions/setup-java@v2
       with:
