+++ .github/workflows/maven-package.yml
@@ -18,7 +18,7 @@ jobs:
     runs-on: macos-latest
 
     steps:
-    - uses: actions/checkout@v2
+    - uses: actions/checkout@v3
     - name: Set up JDK 17
       uses: actions/setup-java@v2
       with:
