+++ .github/workflows/check-standard.yaml
@@ -35,11 +35,7 @@ jobs:
       R_KEEP_PKG_SOURCE: yes
 
     steps:
-      - name: Checkout code
-        uses: rohmanngmbh/action-checkout-repo@v1.3.2
-        with:
-          lfs: true
-          path: RCaNmodel
+      - uses: actions/checkout@v3
 
       - uses: r-lib/actions/setup-pandoc@v2
 
