+++ .github/workflows/maven-build-installer-windows.yml
@@ -29,11 +29,7 @@ jobs:
         pathTarget: ./target/wix
     - name: Add Wix to Path
       run: echo "$HOME/target/wix" >> $GITHUB_PATH
-    - name: Checkout code
-      uses: rohmanngmbh/action-checkout-repo@v1.3.2
-      with:
-        lfs: true
-        path: RCaNconstructor
+    - uses: actions/checkout@v3
     - name: Set up JDK 17
       uses: actions/setup-java@v2
       with:
