### 2023-02-01 #4955 @elfunesto
@fabian-s , I have regenerated the proof, merged your pull request, made a release with tag v2.0.2, archived the repository zenodo with edited metadata and DOI 10.5281/zenodo.7595076
