+++ .github/workflows/maven-build-installer-windows.yml
@@ -45,4 +45,4 @@ jobs:
         automatic_release_tag: "Windows-latest"
         prerelease: true
         title: "Windows Development Build"
-        files: ./target/*.msi
+        files: ./RCaNconstructor/target/*.msi
