+++ .github/workflows/maven-build-installer-unix.yml
@@ -32,4 +32,4 @@ jobs:
         automatic_release_tag: "Ubuntu-latest"
         prerelease: true
         title: "Ubuntu Development Build"
-        files: ./target/*.deb
+        files: ./RCaNconstructor/target/*.deb
