### 2023-02-02 #4955 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
