### 2023-02-02 #4955 @Kevin-Mattheus-Moerman
@elfunesto I am the AEiC for this track and here to help process this work for acceptance in JOSS. I've just reviewed the archive and will review the paper shortly as well.

On the ZENODO archive:
- [ ] Please manually edit the license so that it matches the software license (it currently says `Other(Open)`)
