### 2023-02-03 #4955 @elfunesto
@Kevin-Mattheus-Moerman I added the license zenodo (sorry for missing this: the zenodo tool to select a license is not very user friendly). Regarding the acknowledgment section, we do not need one, there is no specific funding for this software.
